update code (#1)
* Send a gratuitous arp for new vips

existing arp_notify doesn't send gratuitous arp for new ip address added
to an interface. This change will trigger a gratuitous arp for any new
ipv4 ip added to a interface to refresh arp entries in the LAN

Change-Id: Idc440aee64a28482eec72e401c8ec2eca99d0f0d

* stop using config from peer

Config from peer is not implemented but set as config source for backup.

Change-Id: I63b8ebdc5d0eafd389a4d9ba188c90c7aa806538

* fix newline in seesaw_cli

because terminal is switched to raw in seesaw_cli, fmt.Print doesn't do a full newline. This change fixes it by restoring the terminal state when running commands.

Change-Id: I1e65756f74800e366558ad2547da1b5d02e8c23e

* apply timeout to connection in config server fetch

* add timeout to syncClient

The lock is held when dialing which blocks all other functions on
syncClient. Adding a timeout can remit it a bit.

Co-authored-by: Yuan Liu <liuy139@gmail.com>